requirements.txt
absl-py==2.0.0
alabaster==0.7.12
alembic==1.7.7
altgraph==0.17.4
anyio==3.6.1
apiai==1.2.3
apispec==3.3.2
appdirs==1.4.4
argcomplete==2.0.0
astroid==2.5.1
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
audioread==2.1.9
Automat==22.10.0
autopep8==1.5.5
Babel==2.10.1
backcall==0.2.0
bcrypt==4.1.1
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.3.0
blinker==1.4
blis==0.7.11
cached-property==1.5.2
cachelib==0.7.0
cachetools==4.2.1
catalogue==2.0.10
cattrs==1.10.0
certifi==2020.12.5
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
cloudpickle==1.6.0
colorama==0.4.6
coloredlogs==15.0
colorlog==4.8.0
comm==0.2.0
commonmark==0.9.1
comtypes==1.1.10
confection==0.0.4
connexion==2.13.1
constantly==15.1.0
contourpy==1.2.0
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==41.0.5
cssselect==1.2.0
custom-inherit==2.3.2
cycler==0.10.0
cymem==2.0.7
debugpy==1.8.0
decorator==4.4.2
deepface==0.0.75
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.1.1
dnspython==2.2.1
docopt==0.6.2
docutils==0.18.1
dtcwt==0.12.0
EasyProcess==1.1
email-validator==1.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoint2==1.1
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
ffmpeg==1.4
filelock==3.13.1
fire==0.5.0
flake8==3.8.4
Flask==2.3.3
flatbuffers==2.0
fonttools==4.45.1
ftfy==6.1.1
future==0.18.2
fuzzy-c-means==0.0.4
gast==0.3.3
gdown==4.6.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.59.2
gTTS==2.2.3
gunicorn==20.1.0
h11==0.12.0
h5py==3.10.0
helpdev==0.7.1
httpcore==0.14.7
httpx==0.22.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
imageio==2.13.5
imagesize==1.2.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
insta-scrape==2.1.2
install==1.3.5
instaloader==4.8.1
instascrape==0.1
intervaltree==3.1.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.12.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.3.2
Js2Py==0.74
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==23.0.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy-object-proxy==1.7.1
libclang==13.0.0
librosa==0.9.0
llvmlite==0.41.1
lockfile==0.12.2
lxml==4.9.3
Mako==1.2.0
Markdown==3.3.7
markovchain==0.2.5
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.6.1
MechanicalSoup==1.2.0
MIDIUtil==1.2.1
mistune==0.8.4
MouseInfo==0.1.3
mpmath==1.2.1
mss==7.0.1
mtcnn==0.1.1
murmurhash==1.0.10
mutagen==1.45.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
nbclient==0.5.13
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nltk==3.8.1
numba==0.58.1
numpy==1.23.5
numpydoc==1.1.0
oauthlib==3.1.0
opencv-python==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==2.1.3
pandocfilters==1.4.3
paramiko==3.3.1
parsel==1.7.0
parso==0.7.0
pathspec==0.11.2
pathy==0.10.1
pefile==2021.5.24
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pipwin==0.5.2
platformdirs==4.0.0
pluggy==0.13.1
pooch==1.3.0
praat-parselmouth==0.4.1
preshed==3.0.8
prison==0.2.1
promise==2.3
prompt-toolkit==3.0.17
Protego==0.2.1
protobuf==3.15.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/FlRE_DEATH/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl#sha256=50c5bd3401bc33b9f58a7c8005841575354a8b4d3293e68ccc4db59d31f18f9a
pycaw==20181226
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.13
PyDispatcher==2.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.7.2
pyloudnorm==0.1.1
pyls-black==0.4.6
pyls-spyder==0.3.2
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pypandoc==1.8.1
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.3
pyreadline==2.1
pyreadline3==3.4.1
PyRect==0.2.0
pyrsistent==0.18.1
pyscreenshot==3.1
pySmartDL==1.3.4
PySocks==1.7.1
PySoundFile==0.9.0.post1
pyspark==3.1.3
pytaglib==2.0.0
pytest==6.1.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-docx==0.8.11
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytweening==1.0.7
pytz==2022.1
pytzdata==2020.1
PyWavelets==1.5.0
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==2.8.1
QtAwesome==1.0.2
qtconsole==5.0.2
QtPy==1.9.0
queuelib==1.6.2
reel==0.0.8
referencing==0.30.2
regex==2022.10.31
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.0
resampy==0.4.2
retina-face==0.0.13
rfc3986==1.5.0
rich==12.4.1
rope==0.18.0
rpds-py==0.12.0
rsa==4.7.1
scikit-learn==1.3.2
scipy==1.11.3
Scrapy==2.8.0
selenium==3.141.0
service-identity==21.1.0
setproctitle==1.2.3
six==1.16.0
smart-open==6.3.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
soupsieve==2.0.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
srsly==2.4.5
stack-data==0.6.3
swagger-ui-bundle==0.0.9
sympy==1.8
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
textdistance==4.2.1
thinc==8.1.12
threadpoolctl==2.1.0
three-merge==0.1.1
tifffile==2021.11.2
tk==0.1.0
tldextract==3.4.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
trio==0.19.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typed-ast==1.5.5
typer==0.7.0
typing_extensions==4.2.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
unicodecsv==0.14.1
urllib3==1.26.18
validate-email==1.3
vexmessage==0.4.0
w3lib==2.1.1
wasabi==1.1.1
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
WTForms==2.3.3
xlrd==2.0.1
yapf==0.31.0
yodel==0.3.0
youtube-dl==2021.4.26
zipp==3.17.0
zope.interface==5.5.2
